3_flood_risk/src/requirements.txt
PyYAML>=6.0
numpy>=1.26
rasterio>=1.3
geopandas>=0.14
shapely>=2.0
matplotlib>=3.8
fiona>=1.9
pyproj>=3.6
whitebox>=2.3
osmnx>=1.9
pystac-client>=0.7
fsspec>=2024.2
geodatasets>=2023.12
folium>=0.16
branca>=0.7
